requirements.txt
streamlit
pandas
requests
dnspython
beautifulsoup4
openpyxl
tqdm
